pyproject.toml
[build-system]
requires = ["setuptools>=67.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "swarmmaker"
version = "0.1.0"
description = "SwarmMaker multi-agent CLI with LangGraph and Rich UI"
readme = "README.md"
authors = [
    {name = "Agent Swarm"}
]
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.5,<3",
    "rich>=13.7",
    "typer>=0.12",
    "langchain>=0.2",
    "langchain-openai>=0.1.7",
    "langgraph>=0.0.32",
    "langsmith>=0.1.25",
    "httpx>=0.27",
    "tenacity>=8.2",
    "python-dateutil>=2.8",
    "python-dotenv>=1.0.0",
]

[project.scripts]
swarmmaker = "swarmmaker.cli:app"

[tool.setuptools.packages.find]
where = ["src"]
